--- app/components/ResultCard/index.ts
@@ -1,1 +1,0 @@
-export * from "./ResultCard";
